Worked on cache tree parser (needs testing)

+++ static-fire/parser/tests/channel_header_parser_test.go
@@ -20,7 +20,7 @@ Delta_X	1.953125E-5		1.953125E-5
 
 	expected := parser.ParsedChannelHeader{
 		ChannelCount: 2,
-		Samples:      []uint64{51200, 51200},
+		Samples:      []int{51200, 51200},
 		Dates:        []string{"2016/08/23", "2016/08/23"},
 		Times:        []string{"10:45:47.0352557312499836422", "10:45:47.0352557312499836422"},
 		YUnitLabels:  []string{"Volts", "g"},
